+++ src/app/shared/remote-config/interactive-config-step/interactive-config-step.component.ts
@@ -8,7 +8,6 @@ import {
   output,
 } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
@@ -28,7 +27,6 @@ import { getDefaultAnswerFromQuestion } from '@app/services';
     MatSelectModule,
     MatSlideToggleModule,
     MatInputModule,
-    MatButtonModule,
     MatIconModule,
     LineBreaksPipe,
     TranslateModule,